mv func getter for conf -> config_getter.cpp
